requirements.txt
wandb
pandas
matplotlib
tqdm
scikit-learn
torch_tb_profiler
h5py